add testing dependency and one sample unit test

+++ app/src/main/java/com/noteaker/sample/domain/model/Note.kt
@@ -5,7 +5,7 @@ import com.noteaker.sample.data.model.NoteEntity
 
 @Immutable
 data class Note(
-    val id: Int = 0,
+    val id: Long = 0,
     val title: String,
     val note: String,
     val lastUpdated: Long = System.currentTimeMillis(),
@@ -13,7 +13,7 @@ data class Note(
 )
 
 fun Note.toEntity(): NoteEntity = NoteEntity(
-    id = if (id == 0) 0L else id.toLong(),
+    id = if (id.toInt() == 0) 0L else id,
     title = title,
     note = note,
     lastUpdated = lastUpdated
